second repair process code from feedback
